+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/binary/compatibility/validator/HubdleConfigBinaryCompatibilityValidatorExtension.kt
@@ -68,6 +68,3 @@ constructor(
 internal val HubdleEnableableExtension.hubdleBinaryCompatibilityValidator:
     HubdleConfigBinaryCompatibilityValidatorExtension
     get() = getHubdleExtension()
-
-private val Project.apiValidationExtension: ApiValidationExtension
-    get() = getHubdleExtension()